Factor out operationHandlers.js
