+++ src/components/UsfmEditor.js
@@ -7,6 +7,7 @@ import usfmjs from "usfm-js";
 import "./UsfmEditor.css";
 import {UsfmRenderingPlugin} from "./UsfmRenderingPlugin"
 import {toUsfmJsonAndSlateJson} from "./jsonTransforms/usfmjsToSlate";
+import {handleOperation} from "./operationHandlers";
 
 /**
  * A WYSIWYG editor component for USFM
@@ -57,69 +58,13 @@ class UsfmEditor extends React.Component {
 
     handleChange = (change) => {
         this.setState(prevState => {
-            let isDirty = false;
-
             for (const op of change.operations) {
                 console.debug(op.type, op);
-                switch (op.type) {
-                    case 'add_mark':
-                    case 'remove_mark':
-                    case 'set_mark':
-                    case 'set_selection':
-                        break;
 
-                    case 'insert_text': {
-                        isDirty = true;
-                        console.debug(op.type, op);
-                        const node = prevState.value.document.getClosestInline(op.path);
-                        const sourceKey = node.data.get("source");
-                        const source = prevState.sourceMap.get(sourceKey);
-                        console.debug("insert_text node", node);
-                        console.debug("insert_text source", source);
-                        const sourceField = source.type === "contentWrapper" ? "content" : "text";
-                        const sourceText = source[sourceField];
-                        source[sourceField] = sourceText.slice(0, op.offset) + op.text + sourceText.slice(op.offset);
-                        console.debug("insert_text updated source", source);
-                        break;
-                    }
-
-                    case 'remove_text': {
-                        isDirty = true;
-                        console.debug(op.type, op);
-                        const node = prevState.value.document.getClosestInline(op.path);
-                        const sourceKey = node.data.get("source");
-                        const source = prevState.sourceMap.get(sourceKey);
-                        console.debug("remove_text node", node);
-                        console.debug("remove_text source", source);
-                        const sourceField = source.type === "contentWrapper" ? "content" : "text";
-                        const sourceText = source[sourceField];
-                        source[sourceField] = sourceText.slice(0, op.offset) + sourceText.slice(op.offset + op.text.length);
-                        console.debug("remove_text updated source", source);
-                        break;
-                    }
-
-                    case 'insert_node':
-                    case 'merge_node':
-                    case 'move_node':
-                    case 'remove_node':
-                    case 'set_node':
-                    case 'split_node': {
-                        isDirty = true;
-                        break;
-                    }
-
-                    case 'set_value': {
-                        isDirty = true;
-                        break;
-                    }
-
-                    default:
-                        console.warn("Unknown operation", op.type);
+                const {isDirty} = handleOperation(op, prevState.value, prevState.sourceMap);
+                if (isDirty) {
+                    this.scheduleOnChange();
                 }
-            }
-
-            if (isDirty) {
-                this.scheduleOnChange();
             }
 
             return {value: change.value, usfmJsDocument: prevState.usfmJsDocument};
